tests/wpt/tests/html/canvas/offscreen/filters/2d.filter.value.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.filter.value</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.filter.value</h1>
<p class="desc">test if ctx.filter works correctly</p>


<script>
var t = async_test("test if ctx.filter works correctly");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assert(ctx.filter == 'none', "ctx.filter == 'none'");
  ctx.filter = 'blur(5px)';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.save();
  ctx.filter = 'none';
  _assert(ctx.filter == 'none', "ctx.filter == 'none'");
  ctx.restore();
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

  ctx.filter = 'blur(10)';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = 'blur 10px';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

  ctx.filter = 'inherit';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = 'initial';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = 'unset';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

  ctx.filter = '';
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = null;
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");
  ctx.filter = undefined;
  _assert(ctx.filter == 'blur(5px)', "ctx.filter == 'blur(5px)'");

  ctx.filter = 'blur(  5px)';
  assert_equals(ctx.filter, 'blur(  5px)');
  t.done();

});
</script>
